refactor: 에러메세지 출력과 정보 메세지 출력 클래스 분리 및 관련 메소드 명 수정
